requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
DateTime==4.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.0.0
pymongo==4.1.1
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.36
toml==0.10.2
tzdata==2022.1
tzlocal==4.2
uuid==1.30
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
zope.interface==5.4.0
